website/requirements/requirements.txt
Django==1.6.6
Markdown==2.5
MySQL-python==1.2.5
Pillow==2.5.3
South==1.0
amqp==1.4.6
analytics==0.6.5
anyjson==0.3.3
app-version==0.2.1
beautifulsoup4==4.3.2
billiard==3.3.0.18
celery==3.1.15
cmsplugin-filer==0.10
cmsplugin-twitter==1.0.4
crispy-forms-foundation==0.3.5
cssselect==0.9.1
cssutils==1.0
django-absolute==0.3
django-admin-shortcuts==1.2.5
django-appconf==0.6
django-authtools==1.0.0
django-badbrowser==1.0.7
django-celery==3.1.15
django-classy-tags==0.5.1
django-cms==3.0.5
django-compressor==1.3
django-countries==2.1.2
django-crispy-forms==1.4.0
django-debug-toolbar==1.2.1
-e git+https://github.com/hzlf/django-easy-badbrowser.git@0fc653338562854d96a2e9b8d793e8c7ad01ec38#egg=django_easy_badbrowser-master
django-email-as-username==1.6.7
django-emailit==0.2.1
django-extensions==1.4.0
django-filer==0.9.7
django-guardian==1.2.4
django-hvad==0.5.0
django-inspectional-registration==0.4.3
django-mobile==0.4.0
django-mptt==0.6.1
django-pagedown==0.1.0
django-polymorphic==0.5.6
django-registration==1.0
django-reversion==1.8.4
django-sekizai==0.7
django-split-settings==0.1.1
django-social-auth>=0.7.23
django-tagging==0.3.2
django-tastypie==0.12.0
django-xmlrpc==0.1.5
djangocms-admin-style==0.2.2
djangocms-file==0.0.2
djangocms-link==1.4.1
djangocms-picture==0.0.4
djangocms-snippet==1.1
djangocms-teaser==0.0.2
djangocms-text-ckeditor==2.3.0
djangocms-video==0.0.2
easy-thumbnails==2.1
gunicorn==19.1.1
html5lib==1.0b1
httpagentparser==1.7.4
jdcal==1.0
kombu==3.0.23
lxml==3.4.0
nydus==0.10.8
oauthlib==0.6.3
openpyxl==2.0.5
pdfdocument==3.0
premailer==2.5.1
pyparsing==2.0.2
python-dateutil==2.2
python-magic==0.4.6
python-mimeparse==0.1.4
pytz==2014.7
raven==5.0.0
redis==2.10.3
-e git+https://github.com/hzlf/django-relatedadminlinks.git@efb52ccb6ada9146c990ae924f31969999f53518#egg=relatedadminlink-dev
reportlab==3.1.8
requests==2.4.1
requests-oauthlib==0.4.1
simplejson==3.6.3
six==1.8.0
sqlparse==0.1.11
tolerance==0.1.2
wsgiref==0.1.2
pygeocoder==1.2.5
django-geoposition==0.2.2